feat(apiv2): implement EditorURLsKeys with a factories registry.
